+++ Source/Krypton Toolkit/Krypton.Toolkit.Suite.Extended.Messagebox/Globals/GlobalDeclarations.cs
@@ -34,6 +34,7 @@ global using System.Media;
 global using System.Reflection;
 global using System.Runtime.InteropServices;
 global using System.Text;
+global using System.Threading;
 global using System.Windows.Forms;
 
 global using Krypton.Toolkit.Suite.Extended.Developer.Utilities;